requirements.txt
fastapi
uvicorn[standard]
celery[redis]
faster-whisper>=1.1.1
pyannote.audio>=3.3.2
pydantic-settings
python-dotenv
python-multipart